pyproject.toml
[project]
name = "presence-qa-classifier"
version = "0.1.0"
description = "Train and deploy classifier that checks is there an answer on the question in the text."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc>=3.65.0",
    "fire>=0.7.1",
    "hydra-core>=1.3.2",
    "mlflow>=3.8.1",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "optimum[onnxruntime]>=2.1.0",
    "unsloth>=2025.12.9",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
]

[tool.black]
line-length = 100
target-version = ['py313']

[tool.isort]
src_paths = ["presence_qa_classifier"]
profile = "black"
line_length = 100
lines_after_imports = 2

[tool.flake8]
max-line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".eggs",
    "*.egg",
    "build",
    "dist",
]
